Network.Wai.Logger

Description

Apache style logger for WAI applications.

An example:

{-# LANGUAGE OverloadedStrings #-}
module Main where

import Data.ByteString.Builder (byteString)
import Control.Monad.IO.Class (liftIO)
import qualified Data.ByteString.Char8 as BS
import Network.HTTP.Types (status200)
import Network.Wai (Application, responseBuilder)
import Network.Wai.Handler.Warp (run)
import Network.Wai.Logger (withStdoutLogger, ApacheLogger)

main :: IO ()
main = withStdoutLogger $ \aplogger ->
    run 3000 $ logApp aplogger

logApp :: ApacheLogger -> Application
logApp aplogger req response = do
    liftIO $ aplogger req status (Just len)
    response $ responseBuilder status hdr msg
  where
    status = status200
    hdr = [("Content-Type", "text/plain")]
    pong = "PONG"
    msg = byteString pong
    len = fromIntegral $ BS.length pong

High level functions

type ApacheLogger = Request -> Status -> Maybe Integer -> IO ()

Apache style logger.

withStdoutLogger :: (ApacheLogger -> IO a) -> IO a

Executing a function which takes ApacheLogger. This ApacheLogger writes log message to stdout. Each buffer (4K bytes) is flushed every second.

type ServerPushLogger = Request -> ByteString -> Integer -> IO ()

HTTP/2 server push logger in Apache style.

Creating a logger

data ApacheLoggerActions

Function set of Apache style logger.

apacheLogger :: ApacheLoggerActions -> ApacheLogger

The Apache logger.

serverpushLogger :: ApacheLoggerActions -> ServerPushLogger

The HTTP/2 server push logger.

logRotator :: ApacheLoggerActions -> IO ()

This is obsoleted. Rotation is done on-demand. So, this is now an empty action.

logRemover :: ApacheLoggerActions -> IO ()

Removing resources relating to Apache logger. E.g. flushing and deallocating internal buffers.

initLoggerUser :: ToLogStr user => Maybe (Request -> Maybe user) -> IPAddrSource -> LogType -> IO FormattedTime -> IO ApacheLoggerActions

Creating ApacheLogger according to LogType.

initLogger :: IPAddrSource -> LogType -> IO FormattedTime -> IO ApacheLoggerActions

Types

data IPAddrSource

Source from which the IP source address of the client is obtained.

Constructors

FromSocket
From the peer address of the HTTP connection.

FromHeader
From X-Real-IP: or X-Forwarded-For: in the HTTP header.

FromFallback
From the peer address if header is not found.

data LogType' a where

Logger Type.

Constructors

LogNone

Fields

:: LogType' LogStr

No logging.

LogStdout

Fields

:: BufSize

-> LogType' LogStr

Logging to stdout. BufSize is a buffer size for each capability.

LogStderr

Fields

:: BufSize

-> LogType' LogStr

Logging to stderr. BufSize is a buffer size for each capability.

LogFileNoRotate

Fields

:: FilePath

-> BufSize

-> LogType' LogStr

Logging to a file. BufSize is a buffer size for each capability.

LogFile

Fields

:: FileLogSpec

-> BufSize

-> LogType' LogStr

Logging to a file. BufSize is a buffer size for each capability. File rotation is done on-demand.

LogFileTimedRotate

Fields

:: TimedFileLogSpec

-> BufSize

-> LogType' LogStr

Logging to a file. BufSize is a buffer size for each capability. Rotation happens based on check specified in TimedFileLogSpec.

LogCallback

Fields

:: forall a. (a -> IO ())

-> IO ()

-> LogType' a

Logging with a log and flush action. run flush after log each message.

type LogType = LogType' LogStr

data FileLogSpec

The spec for logging files

Constructors

FileLogSpec

Fields

log_file :: FilePath

log_file_size :: Integer

Max log file size (in bytes) before requiring rotation.

log_backup_number :: Int

Max number of rotated log files to keep around before overwriting the oldest one.

Utilities

showSockAddr :: SockAddr -> NumericAddress

Convert SockAddr to NumericAddress. If the address is IPv4-embedded IPv6 address, the IPv4 is extracted.

logCheck :: LogType -> IO ()

Backward compability

clockDateCacher :: IO (DateCacheGetter, DateCacheUpdater)

Returning DateCacheGetter and DateCacheUpdater.

Note: Since version 2.1.2, this function uses the auto-update package internally, and therefore the DateCacheUpdater value returned need not be called. To wit, the return value is in fact an empty action.

type ZonedDate = FormattedTime

A type for zoned date.

type DateCacheGetter = IO ZonedDate

Getting cached ZonedDate.

type DateCacheUpdater = IO ()

Updateing cached ZonedDate. This should be called every second. See the source code of withStdoutLogger.